added popup panel upon marker clicked
