+++ singadrive-web-client/styles/CarInfoPanel.module.scss
@@ -1,0 +1,23 @@
+.car-info-panel {
+  position: fixed;
+  top: 50%;  // Center vertically
+  left: 50%;  // Center horizontally
+  transform: translate(-50%, -50%);  // Adjust for the element's own dimensions
+  width: 300px;
+  background-color: white;
+  border: 1px solid #ccc;
+  padding: 20px;
+  z-index: 1000;
+  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1), 0px 8px 16px rgba(0, 0, 0, 0.1);
+
+  .close-button {
+    position: absolute;
+    top: 10px;
+    right: 10px;
+    background: none;
+    border: none;
+    font-size: 18px;
+    cursor: pointer;
+  }
+}
+
